# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FraudWatch API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, blocked, all]
            default: all
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user (block/role)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/analyses:
    get:
      operationId: listAdminAnalyses
      tags: [admin]
      summary: List all analyses
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All analyses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisListResponse"
  /admin/audit-logs:
    get:
      operationId: getAuditLogs
      tags: [admin]
      summary: Get audit logs
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: userId
          schema:
            type: string
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        isBlocked:
          type: boolean
        totalUploads:
          type: integer
        totalAnalyses:
          type: integer
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clerkId, email, name, role, isBlocked, totalUploads, totalAnalyses, createdAt]
...
    AdminUserUpdate:
      type: object
      properties:
        isBlocked:
          type: boolean
        role:
          type: string
          enum: [user, admin]

    # ─── File ─────────────────────────────────────────────────────────────────
...
    ModelMetrics:
      type: object
      properties:
        accuracy:
          type: number
        precision:
          type: number
        recall:
          type: number
        f1Score:
          type: number
        rocAuc:
          type: number
        trainingTimeMs:
          type: integer
        predictionTimeMs:
          type: integer
      required: [accuracy, precision, recall, f1Score, rocAuc, trainingTimeMs, predictionTimeMs]
    RiskBreakdown:
      type: object
      properties:
        critical:
          type: integer
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
      required: [critical, high, medium, low]
    Analysis:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        fileId:
          type: string
        fileName:
          type: string
        modelName:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        totalTransactions:
          type: ["integer", "null"]
        fraudCount:
          type: ["integer", "null"]
        legitimateCount:
          type: ["integer", "null"]
        fraudPercentage:
          type: ["number", "null"]
        riskBreakdown:
          $ref: "#/components/schemas/RiskBreakdown"
        metrics:
          $ref: "#/components/schemas/ModelMetrics"
        aiSummary:
          type: ["string", "null"]
        featureImportance:
          type: ["object", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required: [id, userId, fileId, fileName, modelName, status, createdAt]
    AnalysisListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Analysis"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]

    # ─── Transaction ──────────────────────────────────────────────────────────
...
    AdminUserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    AuditLog:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userEmail:
          type: string
        action:
          type: string
        resource:
          type: string
        resourceId:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, userEmail, action, resource, createdAt]
    AuditLogListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]